First Commit
Contains early/unstable version of transcoder, which enables drag and
drop filesystem integration in Windows to automatically convert mp3's
into mp4's that can play on the Wii U browser.  Also contains a
directory listing skin for IIS that is formatted to display nicely on
the Wii U Gamepad.